Move AuthService in its own package
